Made room size page - segregated files
