+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__
 temp.py
 test/
 ngrok.exe
+.vscode/
